[FIX] fix backTo

+++ Assets/Packages/Meek.NavigationStack/Runtime/Serivce/Navigation/BackToNavigation.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using UnityEngine;
 using UnityEngine.Pool;
 
 namespace Meek.NavigationStack
@@ -45,7 +46,7 @@ namespace Meek.NavigationStack
                     SkipAnimation = _skipAnimation,
                 });
             }
-            else
+            else if (removeScreenList.Count > 1)
             {
                 foreach (var screen in removeScreenList.Skip(1))
                     await _stackNavigationService.RemoveAsync(screen.GetType(), new RemoveContext());
